backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg[binary]>=3.2.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
bcrypt==4.1.2
Pillow==11.3.0
